Deduplicate two similar name formatting mixins
Resolves a lot of duplication between ConfigurableNaming and
ConfigurableNumbering
