+++ lib/rubocop/cop/mixin/configurable_naming.rb
@@ -5,40 +5,12 @@ module RuboCop
     # This module provides functionality for checking if names match the
     # configured EnforcedStyle.
     module ConfigurableNaming
-      include ConfigurableEnforcedStyle
+      include ConfigurableFormatting
 
-      SNAKE_CASE = /^@{0,2}[\da-z_]+[!?=]?$/
-      CAMEL_CASE = /^@{0,2}_?[a-z][\da-zA-Z]+[!?=]?$/
-
-      def check_name(node, name, name_range)
-        return if operator?(name)
-
-        if valid_name?(node, name)
-          correct_style_detected
-        else
-          add_offense(node, name_range, message(style)) do
-            opposite_style_detected
-          end
-        end
-      end
-
-      def valid_name?(node, name)
-        pattern = (style == :snake_case ? SNAKE_CASE : CAMEL_CASE)
-        name.match(pattern) || class_emitter_method?(node, name)
-      end
-
-      # A class emitter method is a singleton method in a class/module, where
-      # the method has the same name as a class defined in the class/module.
-      def class_emitter_method?(node, name)
-        return false unless node.parent && node.defs_type?
-        # a class emitter method may be defined inside `def self.included`,
-        # `def self.extended`, etc.
-        node = node.parent while node.parent.defs_type?
-
-        node.parent.each_child_node(:class).any? do |c|
-          c.loc.name.is?(name.to_s)
-        end
-      end
+      FORMATS = {
+        snake_case: /^@{0,2}[\da-z_]+[!?=]?$/,
+        camelCase:  /^@{0,2}_?[a-z][\da-zA-Z]+[!?=]?$/
+      }.freeze
     end
   end
 end
